Remove gh-pages job
The gh-pages CI job is incorrect. We maintain two separate repositories,
one for the web site source code and one for the built site. We use the
`build.sh` script to build and deploy the site.

The reason for the separation is so that the source code is not tied
to the site hosting. Just because we currently host the site using
github pages there is no reason that we must continue to do so. Keeping
the hosting separate from the source means we can immediately switch
to a new hosting service by just grabbing the built site and moving it
to the new host, no changes to the source files required.

This separation also allows us to merge things into `master` and
manually control when we deploy while still testing each merge in this
repo.