Add a getting_started page
Most devs know how to do the cargo stuff, its good to have but no need
to be front and centre.

Move the basic cargo stuff to its own page so as not to clutter the
intro.

While we move it, change the text to mention that "rand-std" is useful
for playing around because one can generate keys. Production devs should
probably know what they are doing with secrets anyways.